tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Justin <litch@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.8"
lnd-grpc-client = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
